+++ fili-core/src/main/java/com/yahoo/bard/webservice/table/availability/ConcreteAvailability.java
@@ -9,6 +9,7 @@ import com.yahoo.bard.webservice.util.SimplifiedIntervalList;
 
 import java.util.Collections;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Set;
 
 import javax.validation.constraints.NotNull;
@@ -78,6 +79,28 @@ public class ConcreteAvailability implements Availability {
 
     @Override
     public String toString() {
-        return String.format("ConcreteAvailability for table: %s", name.asName());
+        return String.format(
+                "ConcreteAvailability for table: %s with data source names %s",
+                name.asName(),
+                dataSourceNames
+        );
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (obj instanceof ConcreteAvailability) {
+            ConcreteAvailability that = (ConcreteAvailability) obj;
+            return Objects.equals(name.asName(), that.name.asName())
+                    && Objects.equals(dataSourceNames, that.dataSourceNames)
+                    // Since metadata service is mutable, use instance equality to ensure table equality is stable
+                    && metadataService == that.metadataService;
+        }
+        return false;
+    }
+
+    @Override
+    public int hashCode() {
+        // Leave metadataService out of hash because it is mutable
+        return Objects.hash(name.asName(), dataSourceNames);
     }
 }