+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/util/PaginationParameters.java
@@ -36,6 +36,11 @@ public class PaginationParameters {
             1
     );
 
+    /**
+     * Return only the first result.
+     */
+    public static final PaginationParameters ONE_RESULT = new PaginationParameters(1, 1);
+
     private final int perPage;
     private final int page;